+++ examples/plus_initialization.py
@@ -126,7 +126,7 @@ print("Pattern compilation successful")
 print_pattern(pattern)
 
 # set logical observables
-cout_portmap = compiled_canvas.cout_portset
+cout_portmap = compiled_canvas.cout_portset_cube
 coord2logical_group = {0: PatchCoordGlobal3D((0, 0, 2))}
 logical_observables = {i: cout_portmap[coord] for i, coord in coord2logical_group.items()}
 print(f"Using logical observables: {logical_observables}")
@@ -139,8 +139,8 @@ def create_circuit(pattern: Pattern, noise: float) -> stim.Circuit:
     stim_str = stim_compile(
         pattern,
         logical_observables,
-        after_clifford_depolarization=noise,
-        before_measure_flip_probability=noise,
+        p_depol_after_clifford=noise,
+        p_before_meas_flip=noise,
     )
     return stim.Circuit(stim_str)